fix: 코드 분리
